fix: set <expr> only if the action is a key
A mapping with <expr> set and a RHS callback that is not wrapped with
`vim.schedule()` may cause an error. `vim.schedule()` in return,
however, discards `vim.count` and `vim.count1` such that you can apply a
count to certain mappings.

With this commit, <expr> is only set when required such that we can
safely remove `vim.schedule()`. In case that there is in error, one can
still manually wrap the provided callback with `vim.schedule_wrap()`.